made options a util function and added children props
