+++ src/services/LoginRequest.js
@@ -1,22 +1,5 @@
-import Cookies from 'js-cookie';
+import '../utils/options';
 
-
-const getOptions = credentials => {
-
-    const options = {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Authorization': null,
-        'X-CSRFToken': Cookies.get('csrftoken')
-      },
-      credentials: 'include',
-      body: JSON.stringify(credentials),
-    };
-
-    return options
-  }
 
 function loginRequest(credentials) {
 
@@ -24,7 +7,7 @@ function loginRequest(credentials) {
 
         const env = process.env;
 
-        const url = `${env.REACT_APP_HOST}:${env.REACT_APP_PORT}${env.REACT_APP_LOGIN_PATH}`
+        const url = `${env.REACT_APP_HOST}${env.REACT_APP_LOGIN_PATH}`
 
         return fetch(url, options)
         .then(response => response.json())
